agent_tools/bigquery_mcp_server/requirements.txt

google-cloud-bigquery>=3.36.0
python-docx>=1.2.0
mcp
python-dotenv